Init
After experimenting with currency conversions I learned it is easiest
to use integer math and just let everything have a copperValue. This
avoids unit conversions and copper loss due to round errors inherent in
floating point math in JS.

We now have 2 functions, copperValue() and subUnits(). copperValue
calculates the copper value of an amount of coin. subUnits will fill a
“purse” with an optimized grouping of coins of the copperValue supplied.

Also in this commit:
- update dependencies
- update readme and package.json details
- update eslint rules
- unit tests