parser
Added function to parse coin value expressions.
This was needed to support passing expressions such as “1gp” or
“10gp2SP 32cp”.
This parser was then added to the existing copperValue function to
expand it’s functionality from being able to only determine the value
of a single value/denomination pair to unlimited pairs provided as an
expression. This function can also parse natural language for
value/denomination pairs calculating the copper value in a phrase.